[Eliminar el pdf anterior][giovanni.montes]Eliminar pdf anterior para que gs no de error

+++ pantallas/lib/PhpWord/numero_radicado_word.php
@@ -185,7 +185,9 @@ function combinar_documento($ruta_csv, $directorio_out, $ruta_pdf, $idformato) {
 	if(is_dir($directorio_out)) {
 		$comando1 = 'export HOME=/tmp && libreoffice5.1 --headless -print-to-file --outdir ' . $directorio_out . ' ' . $directorio_out . "*" . $extension_doc;
 		$var = shell_exec($comando1);
-		unlink($ruta_pdf . "documento_word.pdf");
+		if(file_exists($ruta_pdf . "documento_word.pdf")) {
+		    unlink($ruta_pdf . "documento_word.pdf");
+		}
         $comando1 = "gs -sDEVICE=pdfwrite -dNOPAUSE -dBATCH -dSAFER -sOutputFile=" . $ruta_pdf . "documento_word.pdf " . $directorio_out . "*.ps";
 		$var = shell_exec($comando2);
 		//$comando2 = 'export HOME=/tmp && libreoffice5.1 --headless --convert-to pdf:writer_pdf_Export --outdir ' . $directorio_out . ' ' . $directorio_out . "*" . $extension_doc;
